+++ src/test/java/io/appform/dropwizard/sharding/DBShardingBundleTestBase.java
@@ -21,6 +21,7 @@ import com.codahale.metrics.health.HealthCheckRegistry;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Maps;
 import io.appform.dropwizard.sharding.config.ShardedHibernateFactory;
+import io.appform.dropwizard.sharding.config.ShardingBundleOptions;
 import io.appform.dropwizard.sharding.dao.RelationalDao;
 import io.appform.dropwizard.sharding.dao.WrapperDao;
 import io.appform.dropwizard.sharding.dao.testdata.OrderDao;
@@ -58,6 +59,9 @@ public abstract class DBShardingBundleTestBase {
     protected static class TestConfig extends Configuration {
         @Getter
         private ShardedHibernateFactory shards = new ShardedHibernateFactory();
+
+        @Getter
+        private ShardingBundleOptions shardingOptions = new ShardingBundleOptions(true);
     }
 
     protected final TestConfig testConfig = new TestConfig();
